+++ examples/scripts/SPMe.py
@@ -8,64 +8,16 @@ import numpy as np
 pybamm.set_logging_level("INFO")
 
 # load model
-model = pybamm.lithium_ion.SPM(
-    {
-        "SEI": "ec reaction limited",
-        # "SEI film resistance": "none",
-        # "lithium plating": "irreversible",
-    }
-)
+model = pybamm.lithium_ion.SPMe()
 model.convert_to_format = "python"
 
 # create geometry
 geometry = model.default_geometry
 
 # load parameter values and process model and geometry
-parameter_values = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Mohtat2020)
-parameter_values.update(
-    {
-        "Lithium plating kinetic rate constant [m.s-1]": 1e-10,
-        "Initial plated lithium concentration [mol.m-3]": 0,
-        "Lithium metal partial molar volume [m3.mol-1]": 1.3e-5,
-    },
-    check_already_exists=False,
-)
-param = model.param
-
-Vmin = 3.0
-Vmax = 4.2
-Cn = parameter_values.evaluate(param.C_n_init)
-Cp = parameter_values.evaluate(param.C_p_init)
-n_Li_init = parameter_values.evaluate(param.n_Li_particles_init)
-c_n_max = parameter_values.evaluate(param.c_n_max)
-c_p_max = parameter_values.evaluate(param.c_p_max)
-
-esoh_model = pybamm.lithium_ion.ElectrodeSOH()
-esoh_sim = pybamm.Simulation(esoh_model, parameter_values=parameter_values)
-esoh_sol = esoh_sim.solve(
-    [0],
-    inputs={"V_min": Vmin, "V_max": Vmax, "C_n": Cn, "C_p": Cp, "n_Li": n_Li_init},
-)
-
-for var in esoh_model.variables:
-    print(var, esoh_sol[var].data)
-parameter_values.update(
-    {
-        "Initial concentration in negative electrode [mol.m-3]": esoh_sol["x_100"].data[
-            0
-        ]
-        * c_n_max,
-        "Initial concentration in positive electrode [mol.m-3]": esoh_sol["y_100"].data[
-            0
-        ]
-        * c_p_max,
-        "Lower voltage cut-off [V]": Vmin,
-        "Upper voltage cut-off [V]": Vmax,
-    }
-)
-# parameter_values["Current function [A]"] *= -1
-parameter_values.process_model(model)
-parameter_values.process_geometry(geometry)
+param = model.default_parameter_values
+param.process_model(model)
+param.process_geometry(geometry)
 
 # set mesh
 mesh = pybamm.Mesh(geometry, model.default_submesh_types, model.default_var_pts)
@@ -82,14 +34,14 @@ solution = model.default_solver.solve(model, t_eval)
 plot = pybamm.QuickPlot(
     solution,
     [
-        "Negative electrode SOC",
+        "Negative particle concentration [mol.m-3]",
         "Electrolyte concentration [mol.m-3]",
-        "Positive electrode SOC",
+        "Positive particle concentration [mol.m-3]",
         "Current [A]",
         "Negative electrode potential [V]",
         "Electrolyte potential [V]",
         "Positive electrode potential [V]",
-        ["Measured open circuit voltage [V]", "Terminal voltage [V]"],
+        "Terminal voltage [V]",
     ],
     time_unit="seconds",
     spatial_unit="um",