+++ examples/scripts/cycling_ageing_yang.py
@@ -1,5 +1,8 @@
 import pybamm as pb
 
+# Note: the Yang model is still in active development and results do not
+# match with those reported in the paper
+
 pb.set_logging_level("NOTICE")
 model = pb.lithium_ion.Yang2017()
 
@@ -27,22 +30,22 @@ experiment = pb.Experiment(
             "Rest for 30 minutes",
             "Discharge at 1 C until 2.8 V",
         ),
-        # (
-        #     "Charge at 1 C until 4.2 V",
-        #     "Hold at 4.2 V until C/20",
-        #     "Rest for 30 minutes",
-        #     "Discharge at 2 C until 2.8 V",
-        # ),
-        # (
-        #     "Charge at 1 C until 4.2 V",
-        #     "Hold at 4.2 V until C/20",
-        #     "Rest for 30 minutes",
-        #     "Discharge at 3 C until 2.8 V",
-        # ),
+        (
+            "Charge at 1 C until 4.2 V",
+            "Hold at 4.2 V until C/20",
+            "Rest for 30 minutes",
+            "Discharge at 2 C until 2.8 V",
+        ),
+        (
+            "Charge at 1 C until 4.2 V",
+            "Hold at 4.2 V until C/20",
+            "Rest for 30 minutes",
+            "Discharge at 3 C until 2.8 V",
+        ),
     ]
 )
 sim = pb.Simulation(model, experiment=experiment)
-sim.solve(solver=pb.CasadiSolver(mode="safe"))
+sim.solve(solver=pb.CasadiSolver(mode="fast with events"))
 sim.plot(
     [
         "Current [A]",
@@ -59,7 +62,6 @@ sim.plot(
         [
             "Total lithium lost [mol]",
             "Loss of lithium to negative electrode SEI [mol]",
-            # "Loss of lithium to positive electrode SEI [mol]",
         ],
     ]
 )