+++ examples/scripts/experimental_protocols/gitt.py
@@ -6,7 +6,6 @@ import pybamm
 pybamm.set_logging_level("INFO")
 experiment = pybamm.Experiment(
     [("Discharge at C/20 for 1 hour", "Rest for 1 hour")] * 20,
-    termination="80% capacity",
 )
 model = pybamm.lithium_ion.DFN()
 sim = pybamm.Simulation(model, experiment=experiment, solver=pybamm.CasadiSolver())