+++ tests/unit/test_experiments/test_simulation_with_experiment.py
@@ -144,6 +144,84 @@ class TestSimulationExperiment(unittest.TestCase):
         pybamm.set_logging_level("WARNING")
         self.assertEqual(sim._solution, None)
 
+    def test_run_experiment_termination(self):
+        # with percent
+        experiment = pybamm.Experiment(
+            [
+                (
+                    "Discharge at 1C until 3V",
+                    "Charge at 1C until 4.2 V",
+                    "Hold at 4.2V until C/10",
+                ),
+            ]
+            * 10,
+            termination="99% capacity",
+        )
+        model = pybamm.lithium_ion.SPM({"SEI": "ec reaction limited"})
+        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Chen2020)
+        param["SEI kinetic rate constant [m.s-1]"] = 1e-14
+        sim = pybamm.Simulation(model, experiment=experiment, parameter_values=param)
+        sol = sim.solve(solver=pybamm.CasadiSolver())
+        C = sol.summary_variables["Capacity [A.h]"]
+        np.testing.assert_array_less(np.diff(C), 0)
+        # all but the last value should be above the termination condition
+        np.testing.assert_array_less(0.99 * C[0], C[:-1])
+
+        # with Ah value
+        experiment = pybamm.Experiment(
+            [
+                (
+                    "Discharge at 1C until 3V",
+                    "Charge at 1C until 4.2 V",
+                    "Hold at 4.2V until C/10",
+                ),
+            ]
+            * 10,
+            termination="5.04Ah capacity",
+        )
+        model = pybamm.lithium_ion.SPM({"SEI": "ec reaction limited"})
+        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Chen2020)
+        param["SEI kinetic rate constant [m.s-1]"] = 1e-14
+        sim = pybamm.Simulation(model, experiment=experiment, parameter_values=param)
+        sol = sim.solve(solver=pybamm.CasadiSolver())
+        # all but the last value should be above the termination condition
+        np.testing.assert_array_less(5.04, C[:-1])
+
+    def test_save_at_cycles(self):
+        experiment = pybamm.Experiment(
+            [
+                (
+                    "Discharge at 1C until 3.3V",
+                    "Charge at 1C until 4.1 V",
+                    "Hold at 4.1V until C/10",
+                ),
+            ]
+            * 10,
+        )
+        model = pybamm.lithium_ion.SPM()
+        sim = pybamm.Simulation(model, experiment=experiment)
+        sol = sim.solve(
+            solver=pybamm.CasadiSolver("fast with events"), save_at_cycles=2
+        )
+        # Solution saves "None" for the cycles that are not saved
+        for cycle_num in [2, 4, 6, 8]:
+            self.assertIsNone(sol.cycles[cycle_num])
+        for cycle_num in [0, 1, 3, 5, 7, 9]:
+            self.assertIsNotNone(sol.cycles[cycle_num])
+        # Summary variables are not None
+        self.assertIsNotNone(sol.summary_variables["Capacity [A.h]"])
+
+        sol = sim.solve(
+            solver=pybamm.CasadiSolver("fast with events"), save_at_cycles=[3, 4, 5, 9]
+        )
+        # Note offset by 1 (0th cycle is cycle 1)
+        for cycle_num in [1, 5, 6, 7, 9]:
+            self.assertIsNone(sol.cycles[cycle_num])
+        for cycle_num in [0, 2, 3, 4, 8]:
+            self.assertIsNotNone(sol.cycles[cycle_num])
+        # Summary variables are not None
+        self.assertIsNotNone(sol.summary_variables["Capacity [A.h]"])
+
     def test_inputs(self):
         experiment = pybamm.Experiment(
             ["Discharge at C/2 for 1 hour", "Rest for 1 hour"]